Fixes/membrum gain staging (#244)
* docs(membrum): signal-path audit + holistic gain-staging plan

Outputs of the membrum-signal-path workflow audit (run wf_323d9fd1-072):
33 verified code findings + 18 scientific-correctness verdicts diagnosing
why presets sound the same/boring. Root causes are gain-staging collapse at
the voice output, the modal body buried ~17 dB under the noise/click layers,
and motion params inert on 5/6 bodies -- not the mode-frequency math.

Includes the holistic gain-staging design plan (web-verified against
Faust/STK/Smith modal-synthesis practice and EBU true-peak conventions) that
resolves the coupled cluster H-1/H-2/H-4/C-2/M-2/M-3 + the missing main-bus
limiter as one gain structure, with a test-first implementation sequence.


* fix(membrum): holistic gain-staging batch — linear voice + bus limiter (audit H-1/H-2/H-4/C-2)

Resolves the coupled output-stage defects behind 'presets sound the same and
boring' as ONE gain structure (see PLAN-gain-staging-2026-06-07.md), not
isolated patches.

Step 1 — main-bus true-peak limiter (new TruePeakLimiter, Layer 2):
  look-ahead brickwall at -1 dBTP, 4x-oversampled true-peak detection (zero
  added detection latency), stereo-linked, RT-safe (buffers in prepare()).
  Wired on outL/outR after the coupling return, before master gain; 1 ms
  look-ahead reported via getLatencySamples() for host PDC.

Step 2 (H-1/H-4) — per-voice output is now LINEAR: the always-on softClip
  gain stage (which crushed velocity dynamics and stamped one timbre on every
  preset) is replaced by a transparent hardClip safety rail; level_ is a true
  attenuation. The bus limiter owns the ceiling. FeedbackExciter keeps its
  bank-level passivity clip.

Step 3 (C-2) — body normalisation switched from flat 1/sqrt(N) (count-only,
  amplitude-blind) to 1/getInputGainSum() unit-peak + kBodyHeadroom (-12 dBFS).
  Amplitude-aware, so per-body loudness/attack character survives. Dead
  bodyGainCompensation_/secondaryGainCompensation_ make-up removed.

Step 4 (H-2) — noise/click standalone gains recalibrated (3.0->0.243,
  2.0->0.445) so the layers sit ~-18 dBFS, ~6 dB UNDER the body, instead of
  burying it ~17 dB. The body's Material/Size/StrikePos/damping character is
  audible again.

Tests: new true_peak_limiter_test (ceiling/transparency/latency) and
gain_staging_balance_test (layer-vs-body budget); processor latency test.
diagnose_orch body-only goes peak=1.0 CLIPPING -> 0.557 clean; velocity
dynamics restored (vel127>vel30). Golden re-baselined deliberately and the
[.generate_golden] case hidden so the regression actually guards. dsp_tests
(6750) + membrum_tests (551) green; pluginval strictness-5 clean.

Steps 5-6 (ToneShaper drive makeup, NonlinearCoupling AM-only) and clang-tidy
to follow.


* fix(membrum): self-review fixes for gain-staging batch

- Secondary (shell) bank output gain was set from getInputGainSum() BEFORE
  configureSecondaryBank() ran, so it used the previous note's mode profile.
  Moved the setOutputGain/threshold into the enabled branch, after the shell
  modes are configured for the current note. (No effect on the default path —
  the secondary bank is only active when coupling is enabled.)
- Reset the bus true-peak limiter on setActive(true) so a deactivate/reactivate
  cycle without re-setup doesn't emit ~1 ms of stale look-ahead audio.

Build clean, membrum_tests (551) green.


* fix(membrum): make bus true-peak limiter zero-latency (review W1)

The look-ahead limiter delayed only the main bus and reported plugin latency,
which would misalign the aux/multi-out buses against the main bus by ~1 ms
once the host applied delay compensation (VST3 latency is one value for the
whole instance, but only outL/outR were delayed).

Switch to a zero-latency design: instantaneous-attack / smoothed-release peak
limiting on the 4x-oversampled + raw peak, gain applied to the current sample.
No delay ring, no look-ahead, getLatencySamples() stays 0 -> all output buses
remain sample-aligned and no host PDC is needed. Sample peak is still bounded
exactly (tp >= |sample|); inter-sample peaks are bounded within the detector
accuracy, with the -1 dBTP ceiling providing headroom. The trade-off is a
harder attack, acceptable for percussion.

Removed the Processor::getLatencySamples override (base returns 0). Tests
updated: limiter no-delay assertion, transparency at same index, processor
latency == 0. dsp_tests + membrum_tests (551) green; pluginval strictness-5
clean (Reported latency: 0).


* fix(membrum): harden TruePeakLimiter — block-size chunking + NaN guard (review S1/S2)

S1: processBlock now chunks input in maxBlockSize-sized pieces, so a host
block larger than the prepared maxBlockSize can never overflow the oversampling
scratch buffers. Single-chunk (n <= maxBlockSize) behaviour is unchanged.

S2: non-finite (NaN/Inf) input samples are sanitized to 0 via a bit-pattern
test (correct under -ffast-math) before detection, and currentGain_ is clamped
to [0,1]. A stray NaN reaching the bus can no longer latch the gain state and
mute the output; clean samples after a non-finite burst are limited normally.

Tests: oversized-block (300 samples, maxBlockSize 64) and non-finite-input
(NaN/Inf burst then loud signal) cases added. dsp_tests + membrum_tests (551)
green; pluginval strictness-5 clean.


* docs(membrum): mark audit findings H-1/H-2/H-4/M-7 done in the audit report

Add a remediation-status section + inline done/open tags to the signal-path
audit so it serves as the living tracker across sessions: the gain-staging
batch (H-1, H-2, H-4, M-7) is committed; H-3, M-1..M-6, M-8/M-9, the physics
items, preset re-tuning and clang-tidy remain open.


* docs(membrum): fold gain-staging design into the audit; remove the plan file

The audit report is now the single self-contained reference: added a
'Gain-staging design (as built)' appendix capturing the signal flow, gain
budget, body normalisation, the zero-latency limiter, and the design notes for
the still-open M-2/M-3 steps. Deleted PLAN-gain-staging-2026-06-07.md (its
necessary content now lives in the audit; design rationale remains in git
history on this branch).


* fix(membrum): remove NaN guard/test that broke macOS -ffast-math build

The S2 non-finite-input handling (commit c6489e71) failed the macOS CI: under
-ffast-math (-ffinite-math-only) the compiler ASSUMES floats are finite, so
the test's bit-pattern finiteness check (finiteBits) is constant-folded to
'finite' and REQUIRE_FALSE(finiteBits(kNaN)) fails. The volatile-sink trick
makes a NaN bit pattern but cannot make a finiteness ASSERTION reliable in a
-ffast-math TU; the limiter's header-inlined isFinite would likewise be folded
to a no-op under -ffast-math, so the guard was ineffective on that platform
anyway.

Remove the isFinite input sanitisation and its test. The bus input is already
guaranteed finite upstream (per-voice hardClip bounds every voice to [-1,1];
coupling return is limited), so the guard was defensive-only. Keep S1 (block
chunking) and the gain clamp. dsp_tests (6751) + membrum_tests (551) green;
pluginval strictness-5 clean.


---------

+++ plugins/membrum/tests/unit/diagnose/test_orchestral_distortion.cpp
@@ -110,7 +110,7 @@ TEST_CASE("Orchestral timpani kick: bisect distortion source", "[diagnose_orch]"
                 if (a >= 0.99f) ++samplesNearClip;
             }
         }
-        const float rms = std::sqrt(sumSq / kTotalSamples);
+        const float rms = static_cast<float>(std::sqrt(sumSq / kTotalSamples));
 
         std::printf("[diagnose_orch] %-18s peak=%.4f  rms=%.4f  "
                     "samples>=0.99: %d  %s\n",
